requirements.txt
Django>=3.2,<4.0
djangorestframework>=3.12,<4.0
boto3>=1.17,<2.0
django-storages>=1.11,<2.0
django-cors-headers>=3.7,<4.0